Extract EnvironmentList
